refactor(mapping): retire legacy remote-FatSecret mapping engine
Recipe nutrition (resolve-ingredient) now maps through the same unified
local engine as the magic log via a small adapter; the 2,300-line
map-ingredient.ts, its tests, and the debug ops script are deleted. The
shared FatsecretMappedIngredient type lives in map-ingredient-with-fallback.
